完成根据提供的角色id动态更换系统提示词 根据提供的角色id动态更换角色声音 完善每12个小时定时刷新阿里云智能语音交互token

+++ src/main/java/com/dotlinea/soulecho/service/AssistantService.java
@@ -4,5 +4,5 @@ public interface AssistantService {
 
     String SpeechTranscriberWithMicrophone();
 
-    byte[] SpeechRecognition(String string, Long id, Long CharacterId);
+    byte[] SpeechRecognition(String string, Long id, int CharacterId);
 }
